save du jour, lib jusqua memccpr, split ko, makefile ok, en attente de test et de norme

+++ ft_strncmp.c
@@ -6,13 +6,15 @@
 /*   By: cbeaurai <cbeaurai@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/18 23:20:24 by cbeaurai          #+#    #+#             */
-/*   Updated: 2020/11/18 23:31:55 by cbeaurai         ###   ########.fr       */
+/*   Updated: 2020/11/19 19:09:55 by cbeaurai         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "libft.h"
+
 int	strncmp(const char *s1, const char *s2, size_t n)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (i < n)
